requirements.txt
asgiref==3.9.1
Django==5.2.4
django-unfold==0.63.0
pillow==11.3.0
psycopg2==2.9.10
sqlparse==0.5.3
